Make adding tabs to DetailsPanel easier
