+++ src/main/resources/view/DetailsPanel.fxml
@@ -1,25 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-    <VBox fx:id="upcomingEventsPane">
-        <Label fx:id="upcomingEventsTitle">
-            <padding>
-                <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-            </padding>
-        </Label>
-        <ListView fx:id="upcomingDatesListView" VBox.vgrow="ALWAYS"/>
-    </VBox>
-    <VBox fx:id="personDetailsPane">
-        <Label fx:id="personDetailsTitle">
-            <padding>
-                <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-            </padding>
-        </Label>
-        <ListView fx:id="personDetailsListView" VBox.vgrow="ALWAYS"/>
-    </VBox>
-</VBox>
+<VBox fx:id="tabPlaceholder" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" />
